requirements.txt
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
et-xmlfile==1.1.0
flake8==4.0.1
h11==0.13.0
idna==3.3
isort==5.10.1
-e ./
mccabe==0.6.1
numpy==1.22.2
openpyxl==3.0.9
outcome==1.1.0
pandas==1.4.0
Pillow==9.0.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
